pyproject.toml
[tool.poetry]
name = "repl_python3_comp-1510-midterm-hackathon-team-max"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
ansicolors = "^1.1.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
